orchestrator/requirements.txt
asgiref==3.7.2
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.18.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
fastapi==0.68.2
h11==0.14.0
idna==3.4
isodate==0.6.1
minio==7.1.17
msal==1.25.0
msal-extensions==1.0.0
portalocker==2.8.2
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
starlette==0.14.2
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.15.0
